+++ CHANGELOG.md
@@ -59,3 +59,4 @@
 * OS50 - US1415 - [Gerente Técnico] Eu, gerente técnico, devo concluir a análise para validar a aba conclusão da análise.
 * OS50 - US1119 - [Gerente Técnico] Eu, analista, devo acessar em análise para visualizar solicitação de ajustes/justificativa.
 * OS50 - US1109 - [Gerente Técnico] Eu, gerente técnico, devo acessar a caixa de entrada para visualizar processos que solicitam desvínculo.
+* OS50 - US1219 - Eu, analista geo, devo atender as novas regras para enviar comunicado ao órgão.